requirements.txt
pytest~=8.4.0
playwright==1.52.0

allure-pytest==2.14.2
allure-python-commons==2.14.2
python-dotenv>=1.0.0
Appium-Python-Client>=5.1.1
requests>=2.32.3